pipeline/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.14.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
moviepy==1.0.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.2
pyzmq==18.0.1
qtconsole==4.4.3
redis==3.3.11
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.0.1
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.41.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
win-unicode-console==0.5
wrapt==1.11.2
watchdog
